+++ cli/src/folder/list.rs
@@ -33,7 +33,7 @@ impl Command {
     }
 
     #[tracing::instrument(skip_all, level = "info")]
-    pub async fn execute(&self, pcloud: HttpClient, folder_id: usize) {
+    pub async fn execute(&self, pcloud: HttpClient, folder_id: u64) {
         tracing::info!("listing folder {}", folder_id);
         let params = Params::new(folder_id);
         match pcloud.list_folder(&params).await {
